+++ openquake/hmtk/parsers/source_model/__init__.py
@@ -49,6 +49,6 @@
 Package :mod:`openquake.hmtk.parsers.source_model` contains base classes and
 implementations of source model parsers.
 """
-from openquake.hmtk.parsers.source_model.nrml04_parser import nrmlSourceModelParser
+from openquake.hmtk.parsers.source_model.nrml_parser import nrmlSourceModelParser
 
 #__all__ = [nrmlSourceModelParser]
